Restructured backend directory structure.
