+++ apps/backend/app/adaptors/pdf/services.py
@@ -1,13 +1,111 @@
 from __future__ import annotations
 
+from app.core.config import Settings, get_settings
+from app.adaptors.pdf.models import ParsedDocument, SourceParseRequest, QueuedSourceDocument
+
+from dataclasses import dataclass
 from datetime import UTC, datetime
+from fastapi import HTTPException, status
+from hashlib import sha256
+import httpx
+from pathlib import Path
+from pypdf import PdfReader
 from typing import Any, cast
 from urllib.parse import urlparse
 
-import httpx
 
-from app.core.config import Settings
-from app.models.document import QueuedSourceDocument, SourceParseRequest
+@dataclass(slots=True)
+class FetchResult:
+    source_domain: str
+    mime_type: str
+    file_path: Path
+    content_hash: str
+    size_bytes: int
+    source_http_status: int
+
+
+class SourceFetcher:
+    def __init__(
+        self,
+        settings: Settings,
+        client: httpx.AsyncClient | None = None,
+    ) -> None:
+        self.settings = settings
+        self._client = client
+
+    async def fetch_pdf(self, source_url: str, destination_path: Path) -> FetchResult:
+        timeout = httpx.Timeout(self.settings.fetch_timeout_seconds)
+        destination_path.parent.mkdir(parents=True, exist_ok=True)
+        hasher = sha256()
+        total_bytes = 0
+
+        async with httpx.AsyncClient(timeout=timeout) as owned_client:
+            client = self._client or owned_client
+            try:
+                async with client.stream("GET", source_url) as response:
+                    response.raise_for_status()
+
+                    mime_type = response.headers.get(
+                        "content-type",
+                        "application/pdf",
+                    ).split(";")[0]
+                    if mime_type != "application/pdf":
+                        raise HTTPException(
+                            status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="Only PDF source URLs are supported.",
+                        )
+
+                    with destination_path.open("wb") as downloaded_pdf:
+                        async for chunk in response.aiter_bytes(
+                            self.settings.fetch_chunk_size_bytes
+                        ):
+                            total_bytes += len(chunk)
+                            if total_bytes > self.settings.fetch_max_bytes:
+                                raise HTTPException(
+                                    status_code=status.HTTP_413_CONTENT_TOO_LARGE,
+                                    detail=(
+                                        "Source PDF exceeded the configured max size."
+                                    ),
+                                )
+
+                            downloaded_pdf.write(chunk)
+                            hasher.update(chunk)
+
+                    source_domain = urlparse(source_url).hostname or "unknown"
+                    return FetchResult(
+                        source_domain=source_domain,
+                        mime_type=mime_type,
+                        file_path=destination_path,
+                        content_hash=hasher.hexdigest(),
+                        size_bytes=total_bytes,
+                        source_http_status=response.status_code,
+                    )
+            except Exception:
+                if destination_path.exists():
+                    destination_path.unlink()
+                raise
+
+
+class PdfParser:
+    def parse(
+        self,
+        request: SourceParseRequest,
+        fetch_result: FetchResult,
+    ) -> ParsedDocument:
+        reader = PdfReader(str(fetch_result.file_path))
+        extracted_text = ""
+        excerpt = ""
+
+        return ParsedDocument(
+            title=request.title,
+            source_url=request.source_url,
+            source_domain=fetch_result.source_domain,
+            attribution=request.attribution,
+            content_hash=fetch_result.content_hash,
+            page_count=len(reader.pages),
+            extracted_text=extracted_text,
+            extracted_excerpt=excerpt,
+        )
 
 
 class JobStore:
@@ -321,3 +419,134 @@ class _AsyncClientManager:
     async def __aexit__(self, exc_type: Any, exc: Any, traceback: Any) -> None:
         if self._owned_client is not None:
             await self._owned_client.aclose()
+            
+
+class QueuedDocumentProcessor:
+    async def process(self, job_id: str) -> None:
+        settings = get_settings()
+        job_store = JobStore(settings=settings)
+        queued_job = await job_store.get_job(job_id)
+        if queued_job is None:
+            return
+
+        download_path = settings.download_directory / f"{queued_job.id}.pdf"
+        fetcher = SourceFetcher(settings=settings)
+        parser = PdfParser()
+
+        try:
+            await job_store.mark_fetching(job_id)
+            fetch_result = await fetcher.fetch_pdf(
+                str(queued_job.source_url),
+                download_path,
+            )
+
+            await job_store.mark_parsing(
+                job_id,
+                source_http_status=fetch_result.source_http_status,
+            )
+
+            parsed_document = parser.parse(
+                self._build_parse_request(queued_job),
+                fetch_result,
+            )
+
+            await job_store.mark_completed(
+                job_id,
+                content_hash=parsed_document.content_hash,
+                page_count=parsed_document.page_count,
+                source_http_status=fetch_result.source_http_status,
+            )
+        except HTTPException as error:
+            await job_store.mark_failed(
+                job_id,
+                error_message=error.detail,
+                source_http_status=self._extract_http_status(error),
+            )
+        except httpx.HTTPStatusError as error:
+            await job_store.mark_failed(
+                job_id,
+                error_message=(
+                    f"Source fetch failed with HTTP {error.response.status_code}."
+                ),
+                source_http_status=error.response.status_code,
+            )
+        except Exception as error:
+            await job_store.mark_failed(
+                job_id,
+                error_message=str(error) or "Unexpected processing failure.",
+            )
+        finally:
+            if not settings.keep_downloaded_pdfs:
+                self._cleanup_download(download_path)
+
+    def _build_parse_request(
+        self,
+        queued_job: QueuedSourceDocument,
+    ) -> SourceParseRequest:
+        return SourceParseRequest(
+            title=queued_job.title,
+            source_url=queued_job.source_url,
+            attribution=queued_job.attribution,
+            notes=queued_job.notes,
+        )
+
+    def _cleanup_download(self, download_path: Path) -> None:
+        if download_path.exists():
+            download_path.unlink()
+
+    def _extract_http_status(self, error: HTTPException) -> int | None:
+        return error.status_code if isinstance(error.status_code, int) else None
+    
+
+class DocumentQueue:
+    async def enqueue(self, request: SourceParseRequest) -> QueuedSourceDocument:
+        self._validate_pdf_source_url(str(request.source_url))
+        settings = get_settings()
+        job_store = JobStore(settings=settings)
+        return await job_store.create_queued_job(request)
+
+    async def list_items(self) -> list[QueuedSourceDocument]:
+        settings = get_settings()
+        job_store = JobStore(settings=settings)
+        return self._latest_jobs_by_document(await job_store.list_jobs())
+
+    async def list_recoverable_items(self) -> list[QueuedSourceDocument]:
+        settings = get_settings()
+        job_store = JobStore(settings=settings)
+        return self._latest_jobs_by_document(
+            await job_store.list_jobs(statuses=["queued", "fetching", "parsing"])
+        )
+
+    async def get_item(self, job_id: str) -> QueuedSourceDocument | None:
+        settings = get_settings()
+        job_store = JobStore(settings=settings)
+        return await job_store.get_job(job_id)
+
+    async def delete_item(self, job_id: str) -> bool:
+        settings = get_settings()
+        job_store = JobStore(settings=settings)
+        return await job_store.delete_job(job_id)
+
+    def _validate_pdf_source_url(self, source_url: str) -> None:
+        parsed_url = urlparse(source_url)
+        if not parsed_url.path.lower().endswith(".pdf"):
+            raise ValueError("Source URL must point to a PDF document.")
+
+    def _latest_jobs_by_document(
+        self,
+        jobs: list[QueuedSourceDocument],
+    ) -> list[QueuedSourceDocument]:
+        latest_by_document: dict[str, QueuedSourceDocument] = {}
+        ordered_latest_jobs: list[QueuedSourceDocument] = []
+
+        for job in jobs:
+            if job.document_id in latest_by_document:
+                continue
+
+            latest_by_document[job.document_id] = job
+            ordered_latest_jobs.append(job)
+
+        return ordered_latest_jobs
+
+
+document_queue = DocumentQueue()
