+++ apps/backend/app/main.py
@@ -1,7 +1,8 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from app.api.routes import router
+from app.core.routes import router as core_router
+from app.adaptors.pdf.routes import router as pdf_adaptor_router
 from app.core.config import get_settings
 
 settings = get_settings()
@@ -22,4 +23,5 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-app.include_router(router, prefix="/api/v1")
+app.include_router(core_router, prefix="/api/v1")
+app.include_router(pdf_adaptor_router, prefix="/api/v1")
