--- apps/backend/app/models/__init__.py
@@ -1,1 +1,0 @@
-
